Fixed styling for notes page and added text editor for notes
