+++ client/components/TextEditor.jsx
@@ -1,13 +1,32 @@
-import React from 'react';
-import axios from 'axios';
-import ReactDOM from 'react-dom';
-import { Editor, EditorState } from 'draft-js';
-import 'draft-js/dist/Draft.css';
+import React from "react";
+import { EditorState } from "draft-js";
+import "draft-js/dist/Draft.css";
+import { Editor } from "react-draft-wysiwyg";
+import "react-draft-wysiwyg/dist/react-draft-wysiwyg.css";
 
 export default function MyEditor() {
-  const [editorState, setEditorState] = React.useState(
-    () => EditorState.createEmpty(),
+  const [editorState, setEditorState] = React.useState(() =>
+    EditorState.createEmpty()
   );
 
-  return <Editor editorState={editorState} onChange={setEditorState} />;
+  const editor = React.useRef(null);
+  function focusEditor() {
+    editor.current.focus();
+  }
+
+  return (
+    <div
+      // style={{ border: "1px solid black", minHeight: "6em", cursor: "text" }}
+      // onClick={focusEditor}
+    >
+      <Editor
+        editorState={editorState}
+        toolbarClassName="toolbarClassName"
+        wrapperClassName="wrapperClassName"
+        editorClassName="editorClassName"
+        onEditorStateChange={setEditorState}
+      />;
+    </div>
+  );
 }
+
