+++ package.json
@@ -39,6 +39,7 @@
     "react": "^17.0.2",
     "react-calendar": "^3.4.0",
     "react-dom": "^17.0.2",
+    "react-draft-wysiwyg": "^1.14.7",
     "react-router-dom": "^5.2.0",
     "react-slick": "^0.28.1",
     "slick-carousel": "^1.8.1",
